refactor: change name from RedisService to AuthRedisService
